@@ -1,14 +1,16 @@
 module lincast
 
-go 1.15
+go 1.16
 
 require (
+	github.com/NYTimes/gziphandler v1.1.1
 	github.com/gorilla/mux v1.8.0
 	github.com/joomcode/errorx v1.0.3
-	github.com/markbates/pkger v0.17.1
-	github.com/mattn/go-sqlite3 v1.14.4
+	github.com/mattn/go-sqlite3 v1.14.7 // indirect
 	github.com/mmcdole/gofeed v1.1.0
-	github.com/sirupsen/logrus v1.7.0
-	github.com/stretchr/testify v1.4.0
+	github.com/sirupsen/logrus v1.8.1
+	github.com/stretchr/testify v1.7.0
 	gopkg.in/natefinch/lumberjack.v2 v2.0.0
+	gorm.io/driver/sqlite v1.1.4
+	gorm.io/gorm v1.21.14
 )
