@@ -1,11 +1,12 @@
 package podcasts
 
 import (
+	"reflect"
+	"testing"
+
 	"github.com/joomcode/errorx"
 	assert2 "github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/suite"
-	"reflect"
-	"testing"
 )
 
 type PodcastsTestSuite struct {
@@ -29,23 +30,26 @@ func (s *PodcastsTestSuite) TestGetPodcast() {
 	assert := assert2.New(s.T())
 
 	for _, feed := range s.sampleFeeds {
-		p, err := GetPodcast(feed)
+		p, originalFeed, err := GetPodcastData(feed)
 
 		assert.NoErrorf(err, "the podcast should be created without errors (feed %s)", feed)
 		assert.NotNil(p, "the struct returned should contain the info of the podcast and"+
 			" not be nil (feed %s)", feed)
+
+		assert.NotNil(originalFeed, "the original feed (parsed on a structure gofeed.Feed) should be returned")
 	}
 
 	wrongURL := "something-wrong.com"
-	p, err := GetPodcast(wrongURL)
+	p, _, err := GetPodcastData(wrongURL)
 
 	if assert.Error(err, "if the passed url is incorrect an error should be returned") {
-		assert.True(errorx.IsOfType(err, errorx.IllegalFormat), "the error should be of type IllegalFormat")
+		assert.True(errorx.IsOfType(err, errorx.ExternalError), "the error should be of type "+
+			"errorx.ExternalError")
 	}
 	assert.Nil(p, "the returned struct should be nil")
 
 	wrongURL = "http://localhost:8080"
-	p, err = GetPodcast(wrongURL)
+	p, _, err = GetPodcastData(wrongURL)
 
 	if assert.Error(err, "if there is a problem with the request an error should be returned") {
 		assert.True(errorx.IsOfType(err, errorx.ExternalError), "the returned error should be of type "+
@@ -58,15 +62,16 @@ func (s *PodcastsTestSuite) TestGetEpisodes() {
 	assert := assert2.New(s.T())
 
 	for _, feed := range s.sampleFeeds {
-		p, err := GetPodcast(feed)
+		_, originalFeed, err := GetPodcastData(feed)
 		if err != nil {
-			panic(errorx.Decorate(err, "podcast can't be obtained (feed %s)", feed))
-		}
-		if reflect.ValueOf(p).IsNil() {
-			panic(errorx.Decorate(err, "the returned Podcast struct is nil (feed %s)", feed))
+			panic(errorx.Decorate(err, "the feed '%s' can't be obtained", feed))
 		}
 
-		eps, err := p.GetEpisodes()
+		if reflect.ValueOf(originalFeed).IsNil() {
+			panic(errorx.Decorate(err, "the returned gofeed.Feed struct is nil (feed %s), which is not expected", feed))
+		}
+
+		eps, err := GetEpisodes(originalFeed)
 
 		assert.NoError(err, "episodes should be obtained without errors")
 		assert.True(len(*eps) != 0, "a slice with episodes should be returned")
